Fix travis buiild fails
@milotype Hi! thanksfor your help!
Also, the travis job Log has the following result for the commit e97ecba:
E - com.github.calo001.fondo.appdata.xml:com.github.calo001.fondo:43 
The description contains an URL. This is not allowed, please use the <url/> tag to share links.

I think the best approach is to use the tag url instead of a description. For example in App Center a button is shown when the url tag is present.